fix: loading icon issue & search button issue
